将【电流监控】和【电流测试】功能做为2个编译选项。 1、【电流测试】实现插反报警，离线报警，等功能。

+++ Source/Driver/Driver_ADCmpc3901.h
@@ -12,7 +12,7 @@
 #include "Driver_simulateSpi.h"
 #define GAIN_NUM_ADC 2
 void Init_mcp3901(uchar Pga_sel0, uchar Pga_sel1); //≥ı ºªØMCP3901
-double Read_1CH_adc(uchar ch);
+int32_t Read_1CH_adc(uchar ch);
 
 
 #endif /* SOURCE_DRIVER_DRIVER_ADCMPC3901_H_ */
